+++ src/main/java/com/order/system/repository/OrderRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface OrederRepository extends JpaRepository<Order,Long> {
+public interface OrderRepository extends JpaRepository<Order,Long> {
     public Order findByOrderId(Long id);
     @Query("SELECT   o.orderId  FROM Order o JOIN  o.itemList i WHERE i.itemId = :itemId")
     List<Long> findOrdersByItemId(Long itemId);
