+++ src/main/java/com/order/system/service/ItemService.java
@@ -1,0 +1,18 @@
+package com.order.system.service;
+
+import com.order.system.dto.ItemDTO;
+import com.order.system.dto.OrderByItem;
+import com.order.system.entity.Item;
+
+import java.sql.SQLException;
+import java.util.List;
+import java.util.Set;
+
+public interface ItemService {
+    public Set<ItemDTO> saveItem(Set<ItemDTO> itemDTO, Long id);
+    public List<ItemDTO> getAllListOfOrderedItems();
+    public List<ItemDTO> getAllListOfOrderedItemsById(Long id);
+    public ItemDTO updateExistingOrderItem(Long itemId, Long orderId) throws SQLException;
+    public String deleteOrderedItemById(Long id);
+    public OrderByItem getOrderList(Long itemId);
+}
