+++ src/test/java/com/order/system/service/OrderProcessingServiceTest.java
@@ -3,11 +3,12 @@ package com.order.system.service;
 
 import com.order.system.dto.ItemDTO;
 import com.order.system.dto.OrderByItem;
+import com.order.system.dto.OrderDTO;
 import com.order.system.dto.Status;
 import com.order.system.entity.Item;
 import com.order.system.entity.Order;
 import com.order.system.repository.ItemRepository;
-import com.order.system.repository.OrederRepository;
+import com.order.system.repository.OrderRepository;
 import org.junit.Assert;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
@@ -19,19 +20,14 @@ import org.mockito.MockitoAnnotations;
 import org.springframework.beans.BeanUtils;
 
 import java.sql.SQLException;
-import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
-import java.util.stream.Collectors;
 
 
-import static net.bytebuddy.matcher.ElementMatchers.any;
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.mockito.Mockito.doNothing;
 import static org.mockito.Mockito.when;
-import static org.springframework.test.web.client.ExpectedCount.times;
 
 public class OrderProcessingServiceTest {
 
@@ -40,9 +36,9 @@ public class OrderProcessingServiceTest {
         MockitoAnnotations.openMocks(this);
     }
     @InjectMocks
-    OrderProcessingServiceImpl orderProcessingService;
+    OrderServiceImpl orderProcessingService;
     @Mock
-    OrederRepository orederRepository;
+    OrderRepository orederRepository;
     @Mock
     ItemRepository itemRepository;
 
@@ -50,6 +46,7 @@ public class OrderProcessingServiceTest {
     void testSaveOrderWithItems(){
         MockitoAnnotations.initMocks(this);
         Order order=new Order();
+        OrderDTO orderDTO=new OrderDTO();
         order.setOrderId(Long.valueOf("01"));
         order.setCustomerName("xyz");
         order.setCustomerAddress("abc");
@@ -57,15 +54,18 @@ public class OrderProcessingServiceTest {
         order.setDispatchDate("4566666");
         order.setOrderAmount(666);
         order.setNumberOfItems(23);
+        BeanUtils.copyProperties(order,orderDTO);
         //order.setItemList(List.of());
         when(orederRepository.save(order)).thenReturn(order) ;
-        orderProcessingService.saveOrderWithItems(order);
+        orderProcessingService.saveOrderWithItems(orderDTO);
     }
     @Test
     void testGetOrderWithItems() throws SQLException {
         MockitoAnnotations.initMocks(this);
         Order order=new Order();
+        OrderDTO orderDTO=new OrderDTO();
         List<Order>orderList=new ArrayList<>();
+        List<OrderDTO>orderDTOList=new ArrayList<>();
         order.setOrderId(Long.valueOf("01"));
         order.setCustomerName("xyz");
         order.setCustomerAddress("abc");
@@ -75,6 +75,7 @@ public class OrderProcessingServiceTest {
         order.setNumberOfItems(23);
         // order.setItemList(List.of());
         orderList.add(order);
+        BeanUtils.copyProperties(orderList,orderDTOList);
         when(orederRepository.findAll()).thenReturn(orderList) ;
         orderProcessingService.getOrderWithItems();
 
@@ -106,9 +107,9 @@ public class OrderProcessingServiceTest {
         Order order=new Order();
 
         List<Order>orderList=new ArrayList<>();
-        Long id;
-        id=1L;
-        order.setOrderId(id);
+        Long orderId=1L;
+        Long itemId=2L;
+        order.setOrderId(orderId);
         order.setCustomerName("xyz");
         order.setCustomerAddress("abc");
         // order.setOrderDate(LocalDate)"123445");
@@ -117,9 +118,9 @@ public class OrderProcessingServiceTest {
         order.setNumberOfItems(23);
         // order.setItemList(List.of());
         orderList.add(order);
-        when(orederRepository.findByOrderId(id)).thenReturn(order);
+        when(orederRepository.findByOrderId(orderId)).thenReturn(order);
         when(orederRepository.save(order)).thenReturn(order);
-        orderProcessingService.updateOrderWithItems(order,id);
+        orderProcessingService.updateOrderWithItems(itemId,orderId);
         // Assert.assertNotNull(order);
 
     }
@@ -129,19 +130,12 @@ public class OrderProcessingServiceTest {
         Order order=new Order();
         order=null;
         List<Order>orderList=new ArrayList<>();
-        Long id;
-        id=1L;
-//        order.setOrderId(Long.valueOf("01"));
-//        order.setCustomerName("xyz");
-//        order.setCustomerAddress("abc");
-//        //  order.setOrderDate(LocalDate.parse("20230801"));
-//        order.setDispatchDate("4566666");
-//        order.setOrderAmount(666);
-//        order.setNumberOfItems(23);
-//        order.setItemList(List.of());
+        Long orderId=1L;
+        Long itemId=2L;
+
         orderList.add(order);
         when(orederRepository.save( order)).thenReturn(null);
-        orderProcessingService.updateOrderWithItems(order,id);
+        orderProcessingService.updateOrderWithItems(itemId,orderId);
 
     }
     @Test
@@ -165,138 +159,7 @@ public class OrderProcessingServiceTest {
         orderProcessingService.deleteeOrderWithItems(id);
 
     }
-    @Test
-    void testGetAllListOfOrderedItems() throws SQLException {
-        MockitoAnnotations.initMocks(this);
-        Item item=new Item();
-        List<Item>itemList=new ArrayList<>();
-        item.setItemName("xyz");
-        item.setItemPrice(5.0f);
-        item.setItemId(1L);
-        item.setItemQuantity(3);
-        item.setShippedDate("2023-98-01");
-        itemList.add(item);
-        when(itemRepository.findAll()).thenReturn(itemList) ;
-        orderProcessingService.getAllListOfOrderedItems();
 
-    }
-    @Test
-    void testGetAllListOfOrderedItemsById() throws SQLException {
-        MockitoAnnotations.initMocks(this);
-        Item item=new Item();
-        Long id;
-        id=1L;
-        List<Item> itemList=new ArrayList<>();
-        item.setItemName("xyz");
-        item.setItemPrice(5.0f);
-        item.setItemId(id);
-        item.setItemQuantity(3);
-        item.setShippedDate("2023-98-01");
-        itemList.add(item);
-        when(itemRepository.findAll()).thenReturn(itemList) ;
-        orderProcessingService.getAllListOfOrderedItemsById(id);
-
-    }
-    @Test
-    void testUpdateExistingOrderItem() throws SQLException {
-        MockitoAnnotations.initMocks(this);
-        Order order=new Order();
-        order.setCustomerName("xyz");
-        order.setCustomerAddress("abc");
-        //  order.setOrderDate(LocalDate.parse("20230801"));
-        order.setDispatchDate("4566666");
-        order.setOrderAmount(666);
-        order.setNumberOfItems(23);
-        Long orderId=1L;
-        ItemDTO itemDTO=new ItemDTO();
-        Item item=new Item();
-         Long itemid=3L;
-        Item item1=new Item();
-        Set<Item> itemList=new HashSet<>();
-        item.setItemName("xyz");
-        item.setItemPrice(5.0f);
-        item.setItemId(itemid);
-        item.setItemQuantity(3);
-        item.setShippedDate("2023-98-01");
-        itemList.add(item);
-        order.setItemList(itemList);
-        Mockito.when(orederRepository.findByOrderId(orderId)).thenReturn(order);
-       // Set<Long>ids=itemList.stream().map(x->x.getItemId()).collect(Collectors.toSet());
-       Mockito.when(itemRepository.findByItemId(itemid)).thenReturn(item);
-        itemDTO.setItemId(itemid);
-        itemDTO.setItemPrice(item.getItemPrice());
-        itemDTO.setItemName(item.getItemName());
-       Mockito.when(itemRepository.save(item)).thenReturn(item);
-      // BeanUtils.copyProperties(item,itemDTO);
-
-       orderProcessingService.updateExistingOrderItem(item,orderId);
-
-    }
-    @Test
-    void testUpdateExistingNotinOrder() throws SQLException {
-        MockitoAnnotations.initMocks(this);
-
-        Long orderId=1L;
-        ItemDTO itemDTO=new ItemDTO();
-        Item item=new Item();
-        Long itemid=3L;
-        Item item1=new Item();
-        Set<Item> itemList=new HashSet<>();
-        item.setItemName("xyz");
-        item.setItemPrice(5.0f);
-        item.setItemId(itemid);
-        item.setItemQuantity(3);
-        item.setShippedDate("2023-98-01");
-        itemList.add(item);
-
-       // Mockito.when(itemRepository.save(item)).thenReturn(item);
-        testSaveItem();
-      // Mockito.when(orderProcessingService.saveItem(itemList,orderId));
-      Mockito.when(orderProcessingService.updateExistingOrderItem(item,orderId)).thenReturn(itemDTO);
-
-    }
-    @Test
-    void testDeleteOrderedItemById() throws SQLException {
-        MockitoAnnotations.initMocks(this);
-        Item item=new Item();
-        Long id;
-        id=1L;
-        List<Item> itemList=new ArrayList<>();
-        item.setItemName("xyz");
-        item.setItemPrice(5.0f);
-        item.setItemId(id);
-        item.setItemQuantity(3);
-        item.setShippedDate("2023-98-01");
-        itemList.add(item);
-        itemRepository.deleteById(id) ;
-        String s1= orderProcessingService.deleteOrderedItemById(id);
-        Assertions.assertNotNull(s1);
-
-    }
-    @Test
-    void testSaveItem() throws SQLException {
-        MockitoAnnotations.initMocks(this);
-        Order order=new Order();
-        order.setCustomerName("xyz");
-        order.setCustomerAddress("abc");
-        //  order.setOrderDate(LocalDate.parse("20230801"));
-        order.setDispatchDate("4566666");
-        order.setOrderAmount(666);
-        order.setNumberOfItems(23);
-        Long id=1L;
-        Item item=new Item();
-        Set<Item>itemList=new HashSet<>();
-        item.setItemName("xyz");
-        item.setItemPrice(5.0f);
-        item.setItemId(1L);
-        item.setItemQuantity(3);
-        item.setShippedDate("2023-98-01");
-        itemList.add(item);
-        Mockito.when(orederRepository.findByOrderId(id)).thenReturn(order);
-       Mockito.when(itemRepository.save(item)).thenReturn(item) ;
-       Assert.assertNotNull(orderProcessingService.saveItem(itemList,id));
-
-    }
     @Test
     void testGetOrderList() throws SQLException {
         MockitoAnnotations.initMocks(this);
